feat: input validation

+++ src/controllers/problem.controller.js
@@ -1,9 +1,10 @@
 import Problem from "../models/Problem.js";
 import cache from "../services/cache.js";
+import { sanitize } from "../utils.js";
 
 const getAll = async (req, res) => {
-  const limit = req.query.limit ? parseInt(req.query.limit) : 10;
-  const page = req.query.page ? parseInt(req.query.page) : 1;
+  const limit = sanitize(req.query.limit, "number") || 10;
+  const page = sanitize(req.query.page, "number") || 1;
 
   const key = `problems:${limit}:${page}`;
 
@@ -25,7 +26,10 @@ const getAll = async (req, res) => {
 }
 
 const getById = async (req, res) => {
-  const id = req.params.id;
+  const id = sanitize(req.params.id, "mongo");
+  if (!id) {
+    return res.status(400).json({ message: "Missing path id" });
+  }
 
   const key = `problem:${id}`;
 
@@ -38,11 +42,11 @@ const getById = async (req, res) => {
     }
 
     const problem = await Problem.findById(id);
-    await cache.set(key, JSON.stringify(problem), "EX", 600);
-
     if (!problem) {
       return res.status(404).json({ message: "Problem not found" });
     }
+
+    await cache.set(key, JSON.stringify(problem), "EX", 600);
 
     return res.status(200).json(problem);
   } catch (err) {
@@ -77,7 +81,11 @@ const create = async (req, res) => {
 }
 
 const update = async (req, res) => {
-  const id = req.params.id;
+  const id = sanitize(req.params.id, "mongo");
+  if (!id) {
+    return res.status(400).json({ message: "Missing path id" });
+  }
+
   const { title, description, difficulty, tags, testCases } = req.body;
 
   const request = {
@@ -111,7 +119,10 @@ const update = async (req, res) => {
 };
 
 const remove = async (req, res) => {
-  const id = req.params.id;
+  const id = sanitize(req.params.id, "mongo");
+  if (!id) {
+    return res.status(400).json({ message: "Missing path id" });
+  }
 
   const problem = await Problem.findByIdAndDelete(id);
 
@@ -119,6 +130,7 @@ const remove = async (req, res) => {
     return res.status(404).json({ message: "Problem not found" });
   }
 
+  await cache.del(`problem:${id}`);
   return res.status(204).send();
 }
